backend/inventory/requirements.txt
aioredis==2.0.1
anyio==3.6.2
async-timeout==4.0.2
cleo==1.0.0a5
click==8.1.3
crashtest==0.3.1
Deprecated==1.2.13
fastapi==0.85.2
h11==0.14.0
hiredis==2.0.0
httptools==0.5.0
idna==3.4
more-itertools==8.14.0
packaging==21.3
pptree==3.1
pydantic==1.10.2
pylev==1.4.0
pyparsing==3.0.9
python-dotenv==0.21.0
python-ulid==1.1.0
PyYAML==6.0
redis==4.3.4
redis-om==0.1.0
sniffio==1.3.0
starlette==0.20.4
types-redis==4.3.21.3
typing_extensions==4.4.0
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.0
websockets==10.4
wrapt==1.14.1
